stiv_control/requirements.txt
torch
torchvision
numpy
Pillow
pandas
tqdm
